+++ CHANGELOG.md
@@ -9,11 +9,13 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 ### Added
 - `range` field for the forms.
 - `singleSubmit` attribute on all fields to allow only one submit per form to be used as calculation form.
-- "Result output" custom post type.
+- `Result output` custom post type.
 - Blocks for the result output.
 - Calculator form type and necessary filters.
 - Forms can now use `single submit` option to send data without submit button.
 - Setting for single form to hide global msg on submit success.
+- `esFormsRangeCurrent` shortcode to output the current value of the range field.
+- `esFormsROIP` shortcode to output the result output part.
 
 ### Changed
 - `Input` fields now output correct types for e-mail and URL fields, so the experience on mobile devices should be much better.